requirements.txt
beautifulsoup4==4.9.3
bunch==1.0.1
Jinja2==2.10
Pillow==8.0.1
pylzma==0.5.0
PyYAML==5.3.1
requests==2.18.4
Unidecode==1.1.1
